Consolidate code to new package.
