+++ junit5/src/test/java/com/baeldung/param/Person.java
@@ -1,4 +1,4 @@
-package com.baeldung.helpers;
+package com.baeldung.param;
 
 /**
  * Very simple Person entity.
