+++ junit5/src/test/java/com/baeldung/param/PersonValidator.java
@@ -1,4 +1,4 @@
-package com.baeldung.helpers;
+package com.baeldung.param;
 
 import java.util.Arrays;
